+++ apps/backoffice-v2/src/common/components/atoms/Portal/Portal.tsx
@@ -1,0 +1,8 @@
+import { FunctionComponentWithChildren } from '@ballerine/ui';
+import { createPortal } from 'react-dom';
+
+export const Portal: FunctionComponentWithChildren<{
+  target: Parameters<typeof createPortal>[1];
+}> = ({ children, target }) => {
+  return createPortal(children, target);
+};