+++ examples/report-generation-example/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/report-generation-example
 
+## 0.2.12
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.24
+
+## 0.2.11
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/react-pdf-toolkit@1.2.23
+
 ## 0.2.10
 
 ### Patch Changes