+++ packages/eslint-config-react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ballerine/eslint-config-react
 
+## 2.0.13
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.13
+
+## 2.0.12
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/eslint-config@1.1.12
+
 ## 2.0.11
 
 ### Patch Changes