+++ sdks/workflow-browser-sdk/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @ballerine/workflow-browser-sdk
 
+## 0.6.39
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/common@0.9.28
+  - @ballerine/workflow-core@0.6.39
+
+## 0.6.38
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.27
+  - @ballerine/workflow-core@0.6.38
+
+## 0.6.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.37
+  - @ballerine/common@0.9.26
+
+## 0.6.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.36
+  - @ballerine/common@0.9.25
+
+## 0.6.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.35
+
+## 0.6.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.23
+  - @ballerine/workflow-core@0.6.34
+
 ## 0.6.33
 
 ### Patch Changes