+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,60 @@
 # @ballerine/workflows-service
 
+## 0.7.38
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/common@0.9.28
+  - @ballerine/workflow-core@0.6.39
+  - @ballerine/workflow-node-sdk@0.6.39
+
+## 0.7.37
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.27
+  - @ballerine/workflow-core@0.6.38
+  - @ballerine/workflow-node-sdk@0.6.38
+
+## 0.7.36
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.37
+  - @ballerine/common@0.9.26
+  - @ballerine/workflow-node-sdk@0.6.37
+
+## 0.7.35
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.36
+  - @ballerine/common@0.9.25
+  - @ballerine/workflow-node-sdk@0.6.36
+
+## 0.7.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.35
+  - @ballerine/workflow-node-sdk@0.6.35
+
+## 0.7.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.23
+  - @ballerine/workflow-core@0.6.34
+  - @ballerine/workflow-node-sdk@0.6.34
+
 ## 0.7.32
 
 ### Patch Changes